fix an error in order of operations
previously when an operation was completed via "=" there was a previous
number and a current number but no operation to link them. this is
where the user would be intentionally dismissing the result for a
new calculation. now there is a check for this that sets the
variables properly and the user can go from one completed
calculation to another