allow misc keys to operate on results
previously the misc keys could only do operations on the current number.
but there were cases where the display would show a result but behind
the scenes there would be no current number. when the misc key was
pressed in this situation the actual math behind the operation was not
apparent. now the misc keys can selectively pull a previous number back
to current number and allow them to be operated on